Move impl DataQuery for SpaceViewContents -> space_view_contents.rs
