+++ crates/re_space_view/src/lib.rs
@@ -8,7 +8,9 @@ mod screenshot;
 mod space_view_contents;
 mod unreachable_transform_reason;
 
-pub use data_query::{DataQuery, PropertyResolver};
+pub use data_query::{
+    DataQuery, DataResultHandle, DataResultNode, DataResultTree, PropertyResolver,
+};
 pub use screenshot::ScreenshotMode;
 pub use space_view_contents::{DataBlueprintGroup, SpaceViewContents};
 pub use unreachable_transform_reason::UnreachableTransformReason;
